requirements.txt
uvicorn==0.22.0
fastapi==0.103.2
sqlalchemy==2.0.31
pydantic[email]
psycopg2==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.8
bcrypt==4.0.1
PyJWT==2.8.0
pytest==7.2.0
httpx==0.23.1
pytest-asyncio==0.17.0
pytest-dotenv